Django/main/requirements.txt
Django==3.0.4
django-filter==2.3.0
djangorestframework==3.11.1
